Break out indexing helpers
